Crate som_parser_symbols

This crate serves as the syntactical analyser (parser) for the Simple Object Machine.

This particular version of the parser works with the tokens outputted by the lexical analyser, instead of directly reading text.

Modules

lang

SOM-specific parser combinators.

Structs

AstGenCtxtData

Enums

AstGenCtxtType

AstMethodGenCtxtType

Functions

apply

Applies a parser and returns the output value if the entirety of the input has been parsed successfully.

parse_file

Parses the input of an entire file into an AST.

parse_file_no_universe

Parses the input of an entire file into an AST, without access to the universe (system classes are initialized before the Universe itself, and don’t need access to it)

Type Aliases

AstGenCtxt